Add json files for dishes and allergies and begin coding dish generation ability
The files with listed dishes and allergies were included in the repo so they can be called upon. Also, coding was started for the ability to randomly generate dishes. In this commit, only Monday will randomly generate a dish when the schedule is generated from the schedule modal. The teacup icon was also made to trigger the meal generation function as a temporary aide for development. It will not remain this way once the dish generation ability is complete.