Fix schedule function
Fixed the schedule function so that the date auto-populates with the current day when the schedule modal opens and limits the inputs to today or future dates. The button to generate a menu will become disabled if a user tries to enter a past date or if they do not enter a valid date.