Fix dish repetition
Added code to prevent repetition of dishes. The check for repetition is based on the name of the dish to ensure not only that the same dish is not added twice within a week but also that different dishes with the same name are not repeated within the selected week either.