Menu screens for allergens
The menu generates now after screening for any selected allergens. The allergies checkboxes also will only save their state if the user clicks the button to submit allergies. Otherwise any changes will not persist and will not be included in the allergies to be considered when the next menu generation is performed.